refactor(parser): extract workspace tree builder into workspaceTree.ts
Move DiscoveredFile/DiscoveredFolder, createFileNode, buildWorkspaceTree
and createEmptyFileNode out of parser.ts into src/lib/workspaceTree.ts.
Kept separate from tree.ts so tree.ts stays a pure TreeNode utility
module without a parser dependency. parser.ts keeps temporary
re-exports; the back edge they create disappears when the follow-up
updates importers and drops the re-exports.
